ref: split out a model view from the main controller class
